4245-COUNT-COMMAS-IN-RANGE Stats: Time: 1 ms (83.58%), Space: 42.4 MB (72.73%) - LeetHub

+++ 4245-count-commas-in-range/4245-count-commas-in-range.java
@@ -3,7 +3,7 @@ class Solution {
         
        int ans = 0;
 
-       for(long x = 1000 ; x <= n; x *= 1000){
+       for(int x = 1000 ; x <= n; x *= 1000){
            ans += n - x + 1;
        }
        return ans;
